test(test_nst_algorithm): test algorithm & conditionally mock production vgg model (coverage 97.62%)
