--- src/artificial_artwork/pretrained_model/__init__.py
@@ -1,1 +1,0 @@
-from .model_loader import GraphFactory as graph_factory
