+++ src/artificial_artwork/style_layer_selector.py
@@ -25,6 +25,7 @@ class NSTLayersSelection:
 
     @classmethod
     def from_tuples(cls, layers: List[Tuple[str, float]]):
+        print('FROM_TUPLES:', 'layers ->', layers)
         return NSTLayersSelection([NSTStyleLayer(*layer) for layer in layers])
 
     @property
