+++ src/artificial_artwork/style_model/__init__.py
@@ -1,0 +1,1 @@
+from .graph_factory import GraphFactory as graph_factory
