+++ tests/test_algorithm.py
@@ -14,17 +14,18 @@ def image_file_names():
         'style': 'blue-red_w300-h225.jpg'
     })
 
-@pytest.fixture
-def monkeypatch_model(toy_pre_trained_model, monkeypatch):
-    def patch():
-        import artificial_artwork.pretrained_model.model_loader as ml
-        def toy_callback():
-            return toy_pre_trained_model['parameters_loader']
-        monkeypatch.setattr(
-            ml,
-            'load_default_model_parameters',
-            toy_callback)
-    return patch
+# @pytest.fixture
+# def monkeypatch_model(toy_pre_trained_model, monkeypatch):
+#     def patch():
+#         import artificial_artwork.pretrained_model.model_loader as ml
+#         def toy_callback():
+#             print('------\nTOY LOAD MODEL CALL ---------\n')
+#             return toy_pre_trained_model['parameters_loader']
+#         monkeypatch.setattr(
+#             ml,
+#             'load_default_model_parameters',
+#             toy_callback)
+#     return patch
 
 
 @pytest.fixture
@@ -74,97 +75,101 @@ def create_production_algorithm_runner():
     from artificial_artwork.disk_operations import Disk
     
     noisy_ratio = 0.6
-    def _create_production_algorithm_runner(algorithm, termination_condition_adapter):
+    def _create_production_algorithm_runner(algorithm, termination_condition_adapter, model_design):
         algorithm_runner = NSTAlgorithmRunner.default(
             algorithm,
-            lambda matrix: noisy(matrix, noisy_ratio)
+            lambda matrix: noisy(matrix, noisy_ratio),
+            model_design
         )
 
         algorithm_runner.progress_subject.add(
             termination_condition_adapter,
         )
-        algorithm_runner.peristance_subject.add(
+        algorithm_runner.persistance_subject.add(
             StylingObserver(Disk.save_image, convert_to_uint8)
         )
         return algorithm_runner
     return _create_production_algorithm_runner
 
 
+@pytest.fixture
+def get_algorithm_runner(create_algorithm, create_production_algorithm_runner, pre_trained_model):
+    def _get_algorithm_runner(image_manager, termination_condition_adapter, location):
+        algorithm = create_algorithm(
+            image_manager.content_image,
+            image_manager.style_image,
+            # [(layer_id, style_layers_weight) for layer_id in pre_trained_model['picked_layers']],
+            pre_trained_model.style_layers,
+            # [(toy_pre_trained_model['picked_layers'][0], 1.0)],
+            termination_condition_adapter,
+            location
+        )
+        algorithm_runner = create_production_algorithm_runner(
+            algorithm,
+            termination_condition_adapter,
+            type('ModelDesign', (), {
+                'network_layers': pre_trained_model.network_layers,
+                'parameters_loader': pre_trained_model.parameters_loader
+            })
+        )
+        algorithm_runner.NETWORK_OUTPUT = pre_trained_model.output_layer
+        return algorithm_runner
+    return _get_algorithm_runner
+
+
 def test_nst_runner(
+    get_algorithm_runner,
     image_file_names,
-    toy_pre_trained_model,
-    create_production_algorithm_runner,
-    create_algorithm,
     max_iterations_adapter_factory_method,
     image_manager,
     test_image,
-    monkeypatch,
-    monkeypatch_model,
     tmpdir):
     """Test nst algorithm runner.
 
     Runs a simple 'smoke test' by iterating only 3 times.
     """
-    # if 'AA_VGG_19' not in os.environ:
-    # import artificial_artwork.pretrained_model.model_loader as ml
-    def toy_callback():
-        return toy_pre_trained_model['parameters_loader']
-    # ml.load_default_model_parameters = toy_callback
-    # monkeypatch.setattr(
-    #         artificial_artwork.pretrained_model.model_loader,
-    #         'load_default_model_parameters',
-    #         toy_callback)
-    with patch('artificial_artwork.pretrained_model.model_loader.load_default_model_parameters', new=toy_callback):
-    # monkeypatch_model()
-        ITERATIONS = 3
-
-        image_manager.load_from_disk(test_image(image_file_names.content), 'content')
-        image_manager.load_from_disk(test_image(image_file_names.style), 'style')
-
-        assert image_manager.images_compatible == True
-        
-        termination_condition_adapter = max_iterations_adapter_factory_method(ITERATIONS)
-
-        algorithm = create_algorithm(
-            image_manager.content_image,
-            image_manager.style_image,
-            [(toy_pre_trained_model['picked_layers'][0], 1.0)],
-            termination_condition_adapter,
-            tmpdir
-        )
-
-        algorithm_runner = create_production_algorithm_runner(
-            algorithm,
-            termination_condition_adapter
-        )
-
-        # algorithm_runner.run()
-
- 
-
-def test_running_cli(image_file_names, test_image, monkeypatch_model, tmpdir):
-    """Test the main function of the cli; perform neural style transfer.
-
-    Runs a simple 'smoke test' by iterating only 3 times.
-    """
     import os
-    # if 'AA_VGG_19' not in os.environ:
-    monkeypatch_model()
-    iterations = 3
-    response = runner.invoke(cli, [
-        test_image(image_file_names.content),
-        test_image(image_file_names.style),
-        '--iterations',
-        str(iterations),
-        '--location',
-        tmpdir,
-    ])
-    print(response.output)
-    # print(response.stderr)
-    print(response.exception)
-    print(response.exc_info)
-    assert response.exit_code == 0
+    ITERATIONS = 3
+
+    image_manager.load_from_disk(test_image(image_file_names.content), 'content')
+    image_manager.load_from_disk(test_image(image_file_names.style), 'style')
+
+    assert image_manager.images_compatible == True
     
+    termination_condition_adapter = max_iterations_adapter_factory_method(ITERATIONS)
+
+    algorithm_runner= get_algorithm_runner(image_manager, termination_condition_adapter, tmpdir)
+   
+    algorithm_runner.run()
+
     template_string = image_file_names.content + '+' + image_file_names.style + '-' + '{}' + '.png'
     assert os.path.isfile(os.path.join(tmpdir, template_string.format(1)))
-    assert os.path.isfile(os.path.join(tmpdir, template_string.format(iterations)))
+    assert os.path.isfile(os.path.join(tmpdir, template_string.format(ITERATIONS)))
+
+
+# def test_running_cli(image_file_names, test_image, monkeypatch_model, tmpdir):
+#     """Test the main function of the cli; perform neural style transfer.
+
+#     Runs a simple 'smoke test' by iterating only 3 times.
+#     """
+#     import os
+#     # if 'AA_VGG_19' not in os.environ:
+#     monkeypatch_model()
+#     iterations = 3
+#     response = runner.invoke(cli, [
+#         test_image(image_file_names.content),
+#         test_image(image_file_names.style),
+#         '--iterations',
+#         str(iterations),
+#         '--location',
+#         tmpdir,
+#     ])
+#     print(response.output)
+#     # print(response.stderr)
+#     print(response.exception)
+#     print(response.exc_info)
+#     assert response.exit_code == 0
+    
+#     template_string = image_file_names.content + '+' + image_file_names.style + '-' + '{}' + '.png'
+#     assert os.path.isfile(os.path.join(tmpdir, template_string.format(1)))
+#     assert os.path.isfile(os.path.join(tmpdir, template_string.format(iterations)))
