+++ tests/test_cv_model.py
@@ -6,43 +6,9 @@ from artificial_artwork.pretrained_model.model_loader import get_vgg_19_model_pa
 my_dir = os.path.dirname(os.path.realpath(__file__))
 
 
-PRODUCTION_IMAGE_MODEL = os.environ.get('AA_VGG_19', 'PRETRAINED_MODEL_NOT_FOUND')
-
-
-@pytest.fixture
-def pre_trained_models(toy_pre_trained_model):
-    
-    from artificial_artwork.pretrained_model.model_loader import load_default_model_parameters
-    from artificial_artwork.pretrained_model.vgg_architecture import LAYERS as vgg_layers
-    from artificial_artwork.pretrained_model.image_model import LAYERS as picked_layers
-
-    return {
-        # Production vgg pretrained model
-        'vgg': type('PretrainedModel', (), {
-            'parameters_loader': load_default_model_parameters,
-            'vgg_layers': vgg_layers,
-            'picked_layers': picked_layers,
-        }),
-        # Toy simulated pretrained model for (mock) testing
-        'toy': type('PretrainedModel', (), {
-            'parameters_loader': toy_pre_trained_model['parameters_loader'],
-            'vgg_layers': toy_pre_trained_model['model_layers'],
-            'picked_layers': toy_pre_trained_model['picked_layers'],
-        }),
-    }
-
-@pytest.fixture
-def pre_trained_model(pre_trained_models):
-    import os
-    return {
-        True: pre_trained_models['vgg'],
-        False: pre_trained_models['toy'],
-    }[os.path.isfile(PRODUCTION_IMAGE_MODEL)]
-
-
 @pytest.fixture
 def graph_factory():
-    from artificial_artwork.pretrained_model import graph_factory
+    from artificial_artwork.style_model import graph_factory
     return graph_factory
 
 
@@ -64,11 +30,10 @@ def test_pretrained_model(pre_trained_model, graph_factory):
     for i, name in enumerate(pre_trained_model.vgg_layers):
         assert layers[0][i][0][0][0][0] == name
 
-    from artificial_artwork.pretrained_model.model_loader import NSTModelDesign
+    from artificial_artwork.style_model.model_design import NSTModelDesign
 
-    print('-------------', pre_trained_model.picked_layers)
     graph = graph_factory.create(image_specs, NSTModelDesign(
-        pre_trained_model.picked_layers,
-        model_parameters
+        pre_trained_model.network_layers,
+        lambda: model_parameters
     ))
-    assert set(graph.keys()) == set(['input'] + list(pre_trained_model.picked_layers))
+    assert set(graph.keys()) == set(['input'] + list(pre_trained_model.network_layers))
